requirements.txt
accelerate
attrdict
biopython
biotite
databricks-cli
datasets
dvc
dvc-s3
einops
evaluate
fair-esm
fastparquet
gensim
hydra-core
ipykernel
jsonlines
Logbook
matplotlib
modAL-python
nltk
numpy
pandas
peft
protobuf
pyarrow
PyYAML
rotary-embedding-torch
scikit-learn
scipy
seaborn
sentencepiece
shap
skorch
tensorboard
tokenizers
torch
torchaudio
tqdm
transformers
wandb